# src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: 콘서트 예약 서비스 API (Concert Reservation Service)
  description: 콘서트 예약, 좌석 점유, 대기열, 포인트 관리 및 결제를 지원하는 API 명세서입니다.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버
paths:
...
  /schedules/{scheduleId}/reservations:
    post:
      summary: 좌석 임시 배정 요청
      description: 특정 회차의 좌석을 5분간 임시 점유(HELD)합니다. (대기열 토큰 필수)
      security:
        - QueueTokenAuth: []
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
          description: 콘서트 스케줄(회차) 고유 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - seat_id
              properties:
                user_id:
                  type: string
                  example: "user_123"
                seat_id:
                  type: string
                  example: "seat_10"
      responses:
        '201':
          description: 임시 배정 완료
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          reservation_id:
                            type: string
                            example: "reservation_7654"
                          status:
                            type: string
                            example: "HELD"
                          schedule_id:
                            type: string
                            example: "schedule_456"
                          seat_id:
                            type: string
                            example: "seat_10"
                          hold_expire_time:
                            type: string
                            format: date-time
                            example: "2026-12-01T09:05:00Z"
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/ScheduleNotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'

  # 4-1, 4-2. 포인트 조회 및 충전
...
  /reservations/{reservationId}/payments:
    post:
      summary: 결제 요청
      description: 임시 배정된 좌석에 대한 결제를 진행하고 예약을 확정합니다. (대기열 토큰 필수)
      security:
        - QueueTokenAuth: []
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: integer
                  example: 90000
      responses:
        '201':
          description: 결제 성공 (예약 확정)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          payment_id:
                            type: string
                            example: "payment_1234"
                          amount:
                            type: integer
                            example: 90000
                          status:
                            type: string
                            example: "CONFIRMED"
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          description: 해당 예약을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          $ref: '#/components/responses/GoneError'

  # 6-1, 6-2. 대기열 토큰 발급 및 대기번호 조회
...
components:
...
  securitySchemes:
    QueueTokenAuth:
      type: apiKey
      in: header
      name: Queue-Token
      description: 대기열 진입 시 발급받은 UUID 형태의 토큰 문자열입니다.

  # 데이터 공통 규격 정의
  schemas:
    SuccessEnvelope:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "ERROR_CODE"
            message:
              type: string
              example: "에러 상세 메시지 설명"

    # 스케줄 정보 객체
...
  responses:
    BadRequestError:
      description: 잘못된 요청 파라미터 또는 충전/결제 유효성 검사 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ScheduleNotFoundError:
      description: 존재하지 않는 일정/회차 ID 요청
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: "CONCERT_SCHEDULE_NOT_FOUND"
                  message:
                    type: string
                    example: "해당 콘서트의 일정을 찾을 수 없습니다."
...
    ConflictError:
      description: 리소스 충돌 (이미 선점된 좌석 또는 이미 대기열에 진입한 상태)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    GoneError:
      description: 예약 만료 등 리소스가 유효하지 않게 된 상태
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'